Fix event types
